Remove playwright and browser based tests.
Use new /health url on info app for smoke test. That endpoint now runs actual queries and therefore will show if application is a) running and b) can get data from datasette.

The main test will fail if any of the following are true

- HTTP status not 2xx - call to raise_for_status not caught
- json object does not have "status" key with value "OK"
- dataset count not > 0
- entity count not > 0

It doesn't check for specific numbers of datasets or entities as these change.